requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
Pillow>=8.0.0
PyYAML>=5.4.0
scipy>=1.7.0